FROMGIT: ASoC: dapm: Fix a couple uninitialized ret variables
Smatch complains that these variables could be uninitialized.  The first
one in snd_soc_dai_link_event() is probably a false positive, because
probably we know the lists are not empty.  I would normally ignore the
warning, but GCC complains here as well so I just silenced the warning.
The "ret" in snd_soc_dapm_new_dai() does need to be initialized or it
leads to a bogus dereference in the caller.

Fixes: 3bbf5d34fd4a ("ASoC: dapm: Move error handling to snd_soc_dapm_new_control_unlocked")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 2e558a8127de7b2ed3302f9adcf332ba3feeadb2
 https://github.com/broonie/asoc for-next)

BUG=b:123738217
TEST=Test Audio use cases for CML with full SOF patch series applied.

Change-Id: I5a3a3e4e3f27214fc27401c444ce10deba69ac28
Signed-off-by: smit shah <smit.shah@intel.com>
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>