ASoC: SOF: Use header struct with ext bytes put/get data and add checks
This patch adds to bytes put a check to not exceed the topology
defined bytes control length. The tlv control header is used for
retrieving length instead to directly access the header words.

In bytes get the size field in the header is set the to the value
that is passed to the function minus the header length. It
prevents to write past the user space buffer. The size parameter
originates from topology. Checks for size violations vs. topology
and IPC are added.

The other changes add debug prints to kernel log to see the
actual size limits and do code cleanup.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>