ASoC: SOF: add sof vBE and vFE support
This patch adds the sof virtio BE and FE support.
It will create a dev node for the communication to the userspace.

It also handles the virtio vq event. It dispatches the different
vq kicks to different handlers. In the virtio vq handling, it handles
all the ipc events from vFE.

When there is position update from FW, it will check whether there is
an available entry in notification vq. If yes, send the position update
notification immediately. If there is no available entry, add the position
update event in a list. As soon as the notification vq has an available
entry, send the position update to the vFE.

For vFE, it will create a virtual audio device driver and communite with
vBE audio driver to create PCMs and playback/capture streams.

Signed-off-by: Libin Yang <libin.yang@intel.com>