ASoC: SOF: use suspend_late/resume_early instead of suspend/resume
The normal behaviour for PCI devices is that when the system suspend
occurs while the device is runtime suspended, its runtime_resume
method gets called to bring it back to full power. This seems to
cause race condition between runtime_resume and system suspend
and prevents the system from entering S2.

Also, it is recommended that bus controllers be suspended in the late
suspend phase and resumed in the resume early phase. This ensures that
the runtime pm for the devices is disabled before the suspend_late
is called. Therefore, switch to using suspend_late and resume_early
for the SOF PCI device.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Harshapriya N <harshapriya.n@intel.com>