ASoC: SOF: Intel: hda: pass codec_mask with new pdata structure
No change in functionality, just pass the value with a different
structure.

FIXME: explicit cast used to work-around limitation that SOF pdata
declare the machine structure as const - even when we do assign fields
dynamically. Either we keep it this way and use casts, or we fix the SOF
structure to

	/* machine */
	struct platform_device *pdev_mach;
	union {
		struct snd_soc_acpi_mach *machine;
		struct snd_sof_machine *sof_machine;
	};

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>