FROMGIT: ASoC: da7219: MCLK should be enabled before DAI clocks
For platforms using the Common Clock Framework to control the
codec's DAI clocks, MCLK should be enabled prior to DAI clocks
being turned on. For some platforms the codec is already
provided with an MCLK reference and can therefore control MCLK
itself as it needs to.

To improve functionality MCLK is now added as a parent to the
DAI clocks, if MCLK was provided, so that if they are enabled MCLK
will automatically be enabled as a prerequisite by the CCF.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit a6028cc60aad18d5d7c25d99b5cb8c24399387c3
 https://github.com/broonie/asoc for-next)

BUG=b:123738217
TEST=Test Audio use cases for CML with full SOF patch series applied.

Change-Id: Id26ca3301c83b72dcd209178297f056301215935
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>