ASoC: SOF: WIP: Add support for PM aware PCMs and components.
Add support for components to be active during driver D0i3 and to
add a wakeup source for the driver.

This adds some new tokens that allow the driver to enter D0i3 when
a such is stream has been triggered.

This also adds a wake source flag that informs the driver to set a wake
source on the IPC interrupt.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>