ASoC: SOF: add support for binding widgets with external events
Add support for binding widgets to external dapm events.
Define and implement the DAPM event handler for
keyword detect component. The keyword detect component
is responsible for setting the pipeline params and
triggering all the components in the pipeline upon receiving a
PRE_PMU or a POST_PMD DAPM event indicating that the
audio pipeline it is connected to has been activated.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>