ASoC: SOF: topology: Support newer minor ABI topologies
Add support for topologies with greater ABI minor version numbers. The
kernel will keep on parsing topology and not fail on any unknown DAIs,
components or tuples. It will delete any pipeline that fails to be
instantiated by the FW during the PIPELINE_COMPLETE IPC. i.e. pipeline
is deleted in FW and in driver DAPM.


Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>