[NOT FOR UPSTREAM] CODEOWNERS: Initial version
I am not happy with the level of SOF reviews, so from now on we will
designate mandatory reviews for parts of the code so as to distribute
the work and increase accountability.

CODEOWNERS: updates for admins and SoundWire

Nothing new, just make sure admins get notified for all changes, and
add reviewers for SoundWire stuff.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>