ALSA: hda: fix headphone detection failed
In ASoC, cmd io is diabled when there is no user on
the bus and is enabled when a user is active. But unsolicated
event is diabled in snd_hdac_bus_stop_cmd_io and not
enabled in snd_hdac_bus_init_cmd_io. This makes headphone
detection failed.

This patch Enable unsolicated event when cmd io is enabled.

Tested on whiskylake with legacy hda driver and SOF driver

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>