ASoC: SOF: Intel: hda: fix page fault issue via removing race
There is race between sof driver removing and hda codec jack-detecting
work, which will lead to page fault issue as the latter work is
accessing codec device which could be already removed.

Here add cancellation of jack-detecting work before codecs are removed
to avoid the race and fix the issue.

Bug: https://github.com/thesofproject/linux/issues/1067
Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>