ASoC: SOF: Intel: hda: initial SoundWire machine driver autodetect
For now we have a limited number of machine driver configurations, and
we can detect them based on the link configuration returned after
checking hardware and firmware (BIOS) configurations.

It's likely that in the future we will need to check for _ADR match as
well, and possibly detect which Slave devices actually report as
attached.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>