[SQUASH ME]ASoC: SOF: Intel: hda: resume DSP to D0I0 before taking the tx_mutex
With the addition of the DSP D0I3 feature during S0,
before sending a new IPC(via mailbox) the host must ensure
that the DSP is resumed back to D0I0 prior to sending
the IPC. In the current implementation, this is handled
in the platform-specific snd_sof_dsp_send_msg() op by
calling the snd_sof_dsp_set_power_save() op to set the
DSP state to D0I0.

But, this would result in a deadlock as follows:
tx_mutex lock
snd_sof_send_message()
    snd_sof_dsp_set_power_save()
        tx_mutex lock << DEADLOCK.
Therefore, the DSP must be resumed back to D0I0 first before
acquiring the IPC tx_mutex to avoid the deadlock.

Also, the PM_GATE IPC can be handled even when the DSP is in D0I3.
So, introduce a no_pm version of the sof_ipc_tx_message()
API for use with such IPC's.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>