soundwire: intel: follow the recommended hardware flow to set SHIM_SYNC
The current hard-coded value only works for CNL/WHL/CML platforms, it
needs to be updated for IceLake and TigerLake (change from 24 MHz to
38.4 MHz oscillator).

In addition, the values needs to be programmed before powering-up the
links.

Note that the SYNCPRD value is tied only to the XTAL value and not the
current bus frequency or the frame rate.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>