ASoC: SOF: pcm: correct to use proper hw_constraints
We intended to make the buffer_bytes and period_bytes be multiple times
of the period_size, but not of the caps->period_size_min (the real
period_size can be not times of period_size_min).

As we already have rule to make sure that the buffer_size will be the
"period_size * periods" in ALSA core, what we need here actually is only
to make sure the periods is integer.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>