fixup! ASoC: SOF: inform DSP that driver is going to be removed
Callers of snd_sof_device_remove() do not check the return value,
so e.g. in case the FW has crashed, a kernel oops follows when
trying to unload.

Furthermore, I don't think it's practical to stop unload. If
one tries to debug topology issues, if topology load fails to a
FW error, the CTX_SAVE might fail and you could not unload
the driver anymore. I.e. a full reboot is needed for each try.

With the risk that reboot of the DSP might fail, I think we should
allow unload even if IPC is dead.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Reviewed-by: Marcin Rajwa <marcin.rajwa@linux.intel.com>