ASoC: codec: max98373-sdw: fix a amixer error when codec is suspended
If codec is suspended,  a error of Device or resource busy occurs when
executing amixer. Amixer accesses AMP_PATH_GAIN which is a volatile
register and the regmap is set to cache_only when codec is suspended,
so regmap can't get its data from regmap cache and return -EBUSY.

Actually this register can be set by user and remain unchanged, so we
can get its value from regmap cache. This patch removes volatile attribute
of this register and initializes it to let regmap cache work.

Signed-off-by: Rander Wang <rander.wang@intel.com>
Signed-off-by: Ryan Lee <ryans.lee@maximintegrated.com>