ASoC: rt711-sdca: fix IO transfer timed out when unloading the module.
Due to the SCP_SDCA_INTMASK will be cleared by any reset,
the codec driver wants to set the setting back while the device attached again.
However, the codec driver may go to the suspend mode already at this moment.
The previous version tries to set cache_only=N and bypass the cache, then writes the registers.
There is a problem here that the cache_only feature doesn't set back to 'Y' if the codec suspended.
With this condition, the driver will try to access the registers when unloading the module.
But, the master suspended already and doesn't work for register access properly.
Therefore, this patch uses sdw_write_no_pm instead of regmap_write in the rt711_sdca_update_status().

Signed-off-by: Shuming Fan <shumingf@realtek.com>