ASoC: SOF: Intel: hda: Enable power savings in S0/D0i3
On platforms where iDisp codec is the only HD-Audio codec,
when the DSP is entering D0i3 and the iDisp codec is
powered off, stopping CORB/RIRB DMA and powering
off the iDisp-A link yields additional power savings in S0.
Move the logic for stopping CORB/RIRB DMA and powering
off the link to the hda_dsp_set_D0_state() which will
be called both during system suspend and when the DSP
enters/exits D0i3 in S0.

This needs to be addressed only in SOF and does not impact
the legacy HDA driver as it doesn't support DSP D0i3.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>