ASoC: rt711-sdca: fix IO transfer timed out in kmod test
The driver tries to set the SCP_SDCA_IntMask1/2 registers when the module removes.
At this moment, we need to avoid accessing the registers when the cache_only enables.
Therefore, this patch removes the sdw_write_no_pm when the jack sets to null.
However, this is no concern because the driver disables HID 1/2 event
which is controlled by the codec inside.

Signed-off-by: Shuming Fan <shumingf@realtek.com>
Reviewed-by: Rander Wang <rander.wang@intel.com>
Reviewed-by: Bard Liao <bard.liao@intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@intel.com>