fixup! ASoC: SOF: Add 'non_recoverable' parameter to snd_sof_dsp_panic()
To make sure that we do not miss a DSP panic during a real firmware boot
we need to mark the boot_iteration counter to be HDA_FW_BOOT_ATTEMPTS
indicating the DSP panic handlers that there will be no attempts going to
be made to correct the boot.

This is because the DSP ROM boot and the firmware boot is done under the
same fw_state, under the same step in SOF level.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>