ASoC: SOF: ipc4: Require struct sof_ipc4_msg for message tx and rx
IPC4 related code must use the struct sof_ipc4_msg to pass the message to
be sent to the firmware.
Similarly on message reception, the platform code must make sure that
the ipc4 message container is updated before calling the handler.

For message sending the msg_data is mandatory and must point to the ipc4
message container while the msg_bytes is the size of the optional data
part of the message (data_bytes == ipc4_msg->data_size).

On message reception, platform code is responsible to update the ipc4
struct's header (primary and extension) before calling the handler.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>